Povezano mesto i radnik, popravljene sitne greske

+++ poslovna/src/main/java/poslovna/informatika/poslovna/controller/RadniciController.java
@@ -25,9 +25,9 @@ public class RadniciController {
 	@Autowired
 	private RadniciService radniciService;
 	
-	@Autowired
+/*	@Autowired
 	private MagacinService  magacinService;
-	
+*/	
 	@RequestMapping("/all")
 	public ResponseEntity<List<Radnik>> radnici(){
 		
@@ -40,14 +40,14 @@ public class RadniciController {
 		return new ResponseEntity<List<Radnik>>(radniciService.findByMagacin(null),HttpStatus.OK);
 	}
 	
-	@RequestMapping(value="getMagacini", method = RequestMethod.GET)
+	/*@RequestMapping(value="getMagacini", method = RequestMethod.GET)
 	public ResponseEntity<List<Magacin>> getMagacini() {
 
 		List<Magacin> magacin = magacinService.findAll();
 
 		return new ResponseEntity<>(magacin, HttpStatus.OK);
 	}
-	
+	*/
 	@RequestMapping(method=RequestMethod.POST, consumes="application/json")
 	public ResponseEntity<Radnik> addRadnik(@RequestBody Radnik radnik){
 			
@@ -57,8 +57,19 @@ public class RadniciController {
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Radnik> delete(@PathVariable Long id) {
-		Radnik deleted = radniciService.delete(id);
-		return new ResponseEntity<>(deleted, HttpStatus.OK);
+		List<Radnik> radnici = radniciService.findByMagacin(null);
+		boolean obrisati = false;
+		for(Radnik r:radnici){
+			if(r.getId()==id){
+				obrisati=true;
+			}
+		}
+		if(obrisati){
+			Radnik deleted = radniciService.delete(id);
+			return new ResponseEntity<>(deleted, HttpStatus.OK);
+		}else{
+			return new ResponseEntity<>(null, HttpStatus.OK);
+		}	
 	}
 	
 	@RequestMapping(value="getRadnici", method = RequestMethod.GET)
@@ -80,8 +91,8 @@ public class RadniciController {
 		if(!radnik.getAdresa().equals("")){
 			radnici= radnici.stream().filter(e->e.getAdresa().toUpperCase().contains(radnik.getAdresa().toUpperCase())).collect(Collectors.toList());
 		}
-		if(radnik.getMagacin() != null){
-			radnici= radnici.stream().filter(e->e.getMagacin().getNaziv().equals(radnik.getMagacin().getNaziv())).collect(Collectors.toList());
+		if(radnik.getMesto() != null){
+			radnici= radnici.stream().filter(e->e.getMesto().getNaziv().equals(radnik.getMesto().getNaziv())).collect(Collectors.toList());
 		}
 		
 		return new ResponseEntity<>(radnici, HttpStatus.OK);
@@ -94,7 +105,7 @@ public class RadniciController {
 		radnik.setIme(radnikEdit.getIme());
 		radnik.setPrezime(radnikEdit.getPrezime());
 		radnik.setAdresa(radnikEdit.getAdresa());
-		radnik.setMagacin(radnikEdit.getMagacin());
+		radnik.setMesto(radnikEdit.getMesto());
 		
 
 		radniciService.save(radnik);
